Fixing legacy log4j2 test (not related to this PR)
